backend/requirements.txt
fastapi
uvicorn[standard]
openai-agents
qdrant-client
sqlalchemy
asyncpg
python-dotenv
beautifulsoup4
requests
tiktoken
pydantic
httpx
pytest
pytest-asyncio
python-multipart
pillow